requirements.txt
ttkbootstrap
python-dotenv
reportlab
Pillow
shopify-python-api
requests